// src/app/modules/digitalAssets/digitalAsset.interface.ts
export interface IUpdateDigitalAssets {
  hasNetflix?: boolean;
  hasDisneyPlus?: boolean;
  hasPrimeVideo?: boolean;
  hasViu?: boolean;
  hasAppleTVPlus?: boolean;
  hasSpotify?: boolean;
  hasYouTubePremium?: boolean;
  hasAudible?: boolean;
  hasICloud?: boolean;
  hasGoogleDrive?: boolean;
  hasOneDrive?: boolean;
  hasGrab?: boolean;
  hasFoodpanda?: boolean;
  hasBrokerageETF?: boolean;
  brokerageDetails?: string;
}